Implemented delete.

+++ Classes/Views/main/models/view.php
@@ -12,12 +12,13 @@ if (!isset($viewbag) || !isset($this)) die();
     <a href="?create=true" class="button success">Create</a>
 </p>
 <div class="row medium-up-2">
-    <?php foreach ($viewbag['views'] as $index => $item): ?>
+    <?php foreach ($viewbag['views'] as $item): /**@var SmartModel $item */
+        $index = $item->getPrimaryKeyValue() ?>
         <div class="column">
             <div class="callout">
                 <pre><?= $item ?></pre>
                 <div class="button-group expanded small">
-                    <a href="?delete=<?= $index ?>" class="button alert">Delete</a>
+                    <a href="?delete=<?= $index ?>" class="button alert" onclick="return confirm('Are you sure?');">Delete</a>
                     <a href="?edit=<?= $index ?>" class="button">Edit</a>
                 </div>
                 <div class="clearfix"></div>
